requirements.txt
colorama==0.4.6
httpx==0.28.1
logzero==1.7.0
vdf==3.4
flask==3.0.0
flask-socketio==5.3.6
eventlet==0.36.1
pystray==0.19.4
Pillow==10.0.0
